+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Due to package verification, the latest version below is the unpublished version and the date is meaningless.
 however, it has to be formatted properly to pass verification tests.
 
-## [0.2.9-preview] - 2020-1-1
+## [0.2.9-preview] - 2010-5-17
 
 ### Added